refactor getting ctrl/data
